+++ packages/payment/mobile-client/src/config.ts
@@ -1,0 +1,3 @@
+export default {
+  paymentSuccessDomain: 'https://stage-payment-success.web.app',
+}