+++ packages/payment/mobile-client/src/model.ts
@@ -1,4 +1,4 @@
-import { PaymentProviderID, PaymentStatus } from 'wing-b2c-payment-sdk/model'
+import { PaymentProviderID } from 'wing-b2c-payment-sdk/model'
 
 export interface CanMakePaymentWithProviderRequestParamV1 {
   providerId: PaymentProviderID
@@ -10,23 +10,13 @@ export interface PayCallback {
   onFailed: () => Promise<void>
 }
 
-export interface ShowPaymentProcessingParam {
-  txnId: string
-  with: 'app' | 'web'
-  url: string
-  checkPaymentStatus: () => Promise<PaymentStatus | null>
-}
-export type ShowPaymentProcessing = (param: ShowPaymentProcessingParam) => Promise<PaymentStatus>
-
 export interface MakePaymentRequestParamV1 extends PayCallback {
   providerId: PaymentProviderID
   txnId: string
   paymentUrl: string
-  showPaymentProcessing: ShowPaymentProcessing
 }
 
 export interface PayWithProviderRequestParamV1 {
   txnId: string
   paymentUrl?: string
-  showPaymentProcessing: ShowPaymentProcessing
 }